+++ CHANGELOG.md
@@ -30,6 +30,9 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   ([#690](https://github.com/MyIntervals/emogrifier/pull/690))
 
 ### Fixed
+- Descendant attribute selectors (such as `html input[disabled]`)
+  ([#375](https://github.com/MyIntervals/emogrifier/pull/375),
+  [#709](https://github.com/MyIntervals/emogrifier/pull/709))
 - Attribute selectors with hyphen in attribute name
   ([#284](https://github.com/MyIntervals/emogrifier/issues/284),
   [#540](https://github.com/MyIntervals/emogrifier/pull/540),